structure change for rides
